% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsme.R
\name{jsme-shiny}
\alias{jsme-shiny}
\alias{jsmeOutput}
\alias{renderJsme}
\title{Shiny bindings for jsme}
\usage{
jsmeOutput(outputId, width = "100px", height = "100px")

renderJsme(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a jsme}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using jsme within Shiny
applications and interactive Rmd documents.
}
