% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsme.R
\name{jsme}
\alias{jsme}
\title{Create a JSME widget.}
\usage{
jsme(structure = "", width = "650px", height = "650px")
}
\arguments{
\item{structure}{if character string, this is the initial structure
to be shown in the editor.  If left empty or absent, the editor will
start with a blank screen.  If numeric or logical or list or
data.frame, these will be converted to a character matrix and
passed to the editor as a SMILES.  See `?jsme::setMol` for more
information on the formats that JSME can accept.}

\item{width, height}{integer or character string indicating the width,
height of the widget.  These will be passed to `createWidget` and
the units `px` are assumed (if not provided, the default is "650px").}
}
\value{
If successful, the function returns a list of class
  "htmlwidget" containing the JSME widget, ready for inclusion in
  an Rmarkdown document or Shiny app.
  If an error occurs (e.g. because the structure is invalid),
  the return value is printed to the console and NULL is returned.
}
\description{
JSME (Java Structure Editor for Molecular Editing) is a web app for
editing molecular structures.  This function creates a GUI widget that
can be included in an Rmarkdown document, which will contain the JSME
app.  The app will allow users to interactively create or modify a molecular
structure.  The widget will by default have a width and height of 650px,
but these can be adjusted.
}
